net/nfp: fix flow rule action data
The data type of 'mark_id' field in 'struct nfp_net_cmsg_action'
should be 'uint32_t' rather than 'uint16_t'.

Fixes: 0b9079d24808 ("net/nfp: support flow API for CoreNIC firmware")
Cc: stable@dpdk.org

+++ drivers/net/nfp/nfp_net_cmsg.h
@@ -144,7 +144,7 @@ struct nfp_net_cmsg_action {
 	uint16_t action;
 	uint8_t queue;
 	uint8_t spare;
-	uint16_t mark_id;
+	uint32_t mark_id;
 };
 
 enum nfp_net_cfg_mbox_cmd {
